docs: add pglite crate setup helpers
