+++ Cargo.toml
@@ -1,3 +1,8 @@
+[workspace]
+members = [
+    "crates/pglite"
+]
+
 [package]
 name = "dbschema"
 version = "0.1.0"
@@ -22,24 +27,17 @@ postgres = "0.19"
 log = "0.4"
 env_logger = "0.11"
 url = "2"
-wasmtime = { version = "36", optional = true }
-wasmtime-wasi = { version = "36", optional = true }
 postgres-protocol = { version = "0.6", optional = true }
 bytes = { version = "1", optional = true }
 fallible-iterator = { version = "0.2", optional = true }
+pglite = { path = "crates/pglite", optional = true }
 
 [dev-dependencies]
 tempfile = "3"
 
 [features]
 default = []
-pglite = [
-    "wasmtime",
-    "wasmtime-wasi",
-    "postgres-protocol",
-    "bytes",
-    "fallible-iterator",
-]
+pglite = ["dep:pglite", "postgres-protocol", "bytes", "fallible-iterator"]
 
 [profile.release]
 opt-level = 3
