+++ crates/pglite/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "pglite"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = "1"
+bytes = "1"
+fallible-iterator = "0.2"
+postgres-protocol = "0.6"
+wasmtime = "36"
+wasmtime-wasi = { version = "36", features=["preview1"] }
