+++ crates/pglite/justfile
@@ -1,0 +1,9 @@
+set shell := ["bash", "-cu"]
+root := `git rev-parse --show-toplevel`
+
+# Download the PGlite WASM artifacts
+pglite-assets:
+  set -euo pipefail
+  mkdir -p {{root}}/vendor/pglite
+  curl -L -o {{root}}/vendor/pglite/pglite.wasm https://cdn.jsdelivr.net/npm/@electric-sql/pglite/dist/pglite.wasm
+  curl -L -o {{root}}/vendor/pglite/pglite.data https://cdn.jsdelivr.net/npm/@electric-sql/pglite/dist/pglite.data
